+++ src/repo.rs
@@ -3,12 +3,12 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Namespace {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
     pub path: String,
     pub kind: String,
     pub full_path: String,
-    pub parent_id: Option<i32>,
+    pub parent_id: Option<i64>,
     pub avatar_url: Option<String>,
     pub web_url: String,
 }
@@ -29,7 +29,7 @@ pub struct Links {
 pub struct ContainerExpiryPolicy {
     pub cadence: String,
     pub enabled: bool,
-    pub keep_n: i32,
+    pub keep_n: i64,
     pub older_than: String,
     pub name_regex: String,
     pub name_regex_keep: Option<String>,
@@ -38,7 +38,7 @@ pub struct ContainerExpiryPolicy {
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Owner {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
     pub username: String,
     pub state: String,
@@ -48,16 +48,16 @@ pub struct Owner {
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Group {
-    pub group_id: i32,
+    pub group_id: i64,
     pub group_name: String,
     pub group_full_path: String,
-    pub group_access_level: i32,
+    pub group_access_level: i64,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Access {
-    access_level: i32,
-    notification_level: i32,
+    access_level: i64,
+    notification_level: i64,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
@@ -68,7 +68,7 @@ pub struct Permissions {
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Repository {
-    pub id: i32,
+    pub id: i64,
     pub description: String,
     pub name: String,
     pub name_with_namespace: String,
@@ -82,8 +82,8 @@ pub struct Repository {
     pub web_url: String,
     pub readme_url: Option<String>,
     pub avatar_url: Option<String>,
-    pub forks_count: i32,
-    pub star_count: i32,
+    pub forks_count: i64,
+    pub star_count: i64,
     pub last_activity_at: DateTime<Utc>,
     pub namespace: Namespace,
     pub _links: Links,
@@ -116,13 +116,13 @@ pub struct Repository {
     pub emails_disabled: Option<bool>,
     pub shared_runners_enabled: bool,
     pub lfs_enabled: bool,
-    pub creator_id: i32,
+    pub creator_id: i64,
     pub import_status: String,
-    pub open_issues_count: i32,
-    pub ci_default_git_depth: i32,
+    pub open_issues_count: i64,
+    pub ci_default_git_depth: i64,
     pub ci_forward_deployment_enabled: bool,
     pub public_jobs: bool,
-    pub build_timeout: i32,
+    pub build_timeout: i64,
     pub auto_cancel_pending_pipelines: String,
     pub build_coverage_regex: Option<String>,
     pub ci_config_path: Option<String>,
