+++ src/user.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct User {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
     pub username: String,
     pub state: String,
@@ -22,15 +22,15 @@ pub struct User {
     pub job_title: String,
     pub bot: bool,
     pub work_information: Option<String>,
-    pub followers: i32,
-    pub following: i32,
+    pub followers: i64,
+    pub following: i64,
     pub last_sign_in_at: DateTime<Utc>,
     pub confirmed_at: DateTime<Utc>,
     pub last_activity_on: NaiveDate,
     pub email: String,
-    pub theme_id: i32,
-    pub color_scheme_id: i32,
-    pub projects_limit: i32,
+    pub theme_id: i64,
+    pub color_scheme_id: i64,
+    pub projects_limit: i64,
     pub current_sign_in_at: DateTime<Utc>,
     // TODO
     // pub identities: Vec<()>,
@@ -40,9 +40,9 @@ pub struct User {
     pub external: bool,
     pub private_profile: bool,
     // TODO
-    // pub shared_runners_minutes_limit: Option<i32>,
+    // pub shared_runners_minutes_limit: Option<i64>,
     // TODO
-    // pub extra_shared_runners_minutes_limit: Option<i32>,
+    // pub extra_shared_runners_minutes_limit: Option<i64>,
     pub is_admin: bool,
     pub note: String,
     pub using_license_seat: bool,
